update spatial index library

+++ src/3rdparty/spatialindexnavel/include/spatialindex/capi/DataStream.h
@@ -6,7 +6,7 @@
  * Copyright (c) 2009, Howard Butler
  *
  * All rights reserved.
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation
@@ -25,7 +25,7 @@
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
 ******************************************************************************/
- 
+
 #pragma once
 
 #include "sidx_export.h"
@@ -33,7 +33,7 @@
 class SIDX_DLL DataStream : public SpatialIndex::IDataStream
 {
 public:
-    DataStream(int (*readNext)(SpatialIndex::id_type* id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, uint32_t *nDataLength));
+    DataStream(int (*readNext)(SpatialIndex::id_type* id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, size_t *nDataLength));
     ~DataStream();
 
     SpatialIndex::IData* getNext();
@@ -47,8 +47,8 @@ protected:
     SpatialIndex::id_type m_id;
 
 private:
-    int (*iterfunct)(SpatialIndex::id_type *id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, uint32_t *nDataLength);
-    
+    int (*iterfunct)(SpatialIndex::id_type *id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, size_t *nDataLength);
+
     bool readData();
     bool m_bDoneReading;
