+++ src/3rdparty/spatialindexnavel/include/spatialindex/capi/Index.h
@@ -35,7 +35,7 @@ class SIDX_DLL Index
 
 public:
     Index(const Tools::PropertySet& poProperties);
-    Index(const Tools::PropertySet& poProperties, int (*readNext)(SpatialIndex::id_type *id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, uint32_t *nDataLength));
+    Index(const Tools::PropertySet& poProperties, int (*readNext)(SpatialIndex::id_type *id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, size_t* nDataLength));
     ~Index();
 
     const Tools::PropertySet GetProperties() { index().getIndexProperties(m_properties);  return m_properties;}
